pyproject.toml
[project]
name = "mkdocs-scc"
license = { file = "LICENSE.md" }
readme = "README.md"
requires-python = ">= 3.10"
dynamic = [
  "version"
]

dependencies = [
  "mkdocs-material[recommended,imaging]~=9.5",
  "mkdocs-awesome-pages-plugin~=2.9",
  "mkdocs-git-revision-date-localized-plugin~=1.3",
  "mkdocs-glightbox~=0.4",
  "qrcode~=8.0",
]

[project.optional-dependencies]
insiders = [
  "mkdocs-material[recommended,imaging] @ git+https://github.com/squidfunk/mkdocs-material-insiders.git"
]

[project.entry-points."mkdocs.plugins"]
"scc/blog" = "src.plugins.blog.plugin:CustomBlogPlugin"
"scc/meta" = "src.plugins.meta.plugin:MetaPlugin"
"scc/specs" = "src.plugins.specs.plugin:SpecsPlugin"
